requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.1.0
Flask-Limiter>=3.5.0
Flask-Cors>=4.0.0
Flask-Login>=0.6.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.80.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
cryptography>=40.0.0
python-jose>=3.3.0
email-validator>=2.0.0
PyJWT>=2.7.0
argon2-cffi>=21.0.0
requests>=2.31.0
python-dateutil>=2.8.0
pyotp>=2.8.0
qrcode>=7.4.0
Pillow>=10.0.0
bleach>=6.0.0
